값타입 매핑의 실전 예시
* 값타입 매핑 실전 예시 코딩
* ch4 실전 예시 위에 바로 작업.
