+++ jpaStudy/ch4realex/src/main/java/test/onlineshop/domain/Member.java
@@ -11,9 +11,12 @@ public class Member extends BaseEntity{
 	@Column(name = "MEMBER_ID")
 	private Long id;
 	private String name;
-	private String city;
-	private String street;
-	private String zipcode;
+//	private String city;
+//	private String street;
+//	private String zipcode;
+
+	@Embedded
+	private Address address;
 
 	@OneToMany(mappedBy = "member")
 	private List<Order> orders = new ArrayList<>();
@@ -29,14 +32,14 @@ public class Member extends BaseEntity{
 //		this.zipcode = zipcode;
 //	}
 
-	public Member(Long id, String name, String city, String street, String zipcode, List<Order> orders) {
-		this.id = id;
-		this.name = name;
-		this.city = city;
-		this.street = street;
-		this.zipcode = zipcode;
-		this.orders = orders;
-	}
+//	public Member(Long id, String name, String city, String street, String zipcode, List<Order> orders) {
+//		this.id = id;
+//		this.name = name;
+//		this.city = city;
+//		this.street = street;
+//		this.zipcode = zipcode;
+//		this.orders = orders;
+//	}
 
 	public Long getId() {
 		return id;
@@ -54,27 +57,35 @@ public class Member extends BaseEntity{
 		this.name = name;
 	}
 
-	public String getCity() {
-		return city;
+	public Address getAddress() {
+		return address;
 	}
 
-	public void setCity(String city) {
-		this.city = city;
+	public void setAddress(Address address) {
+		this.address = address;
 	}
 
-	public String getStreet() {
-		return street;
-	}
-
-	public void setStreet(String street) {
-		this.street = street;
-	}
-
-	public String getZipcode() {
-		return zipcode;
-	}
-
-	public void setZipcode(String zipcode) {
-		this.zipcode = zipcode;
-	}
+	//	public String getCity() {
+//		return city;
+//	}
+//
+//	public void setCity(String city) {
+//		this.city = city;
+//	}
+//
+//	public String getStreet() {
+//		return street;
+//	}
+//
+//	public void setStreet(String street) {
+//		this.street = street;
+//	}
+//
+//	public String getZipcode() {
+//		return zipcode;
+//	}
+//
+//	public void setZipcode(String zipcode) {
+//		this.zipcode = zipcode;
+//	}
 }
